# specs/04-api/openapi.yaml
openapi: 3.0.3
info:
  title: RAEED API
  version: "1.0"
  description: >
    Core-flow contract for the RAEED MVP. Covers the endpoints with the most business logic
    (auth, children, sessions/attendance, memories, dashboard) in full; the remaining routes
    follow the same envelope/auth/error conventions documented in conventions.md and are
    listed there as a catalog rather than fully specified here — extend this file as each
    module is implemented, it is the contract both the Flutter app and the React dashboard
    generate typed clients from.
servers:
  - url: https://{host}/api/v1
    variables:
      host:
        default: localhost:3000
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    conversationId:
      name: conversationId
      in: path
      required: true
      schema: { type: string, format: uuid }
    postId:
      name: postId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: attendance.already_recorded }
            message: { type: string }
            details: { type: object, additionalProperties: true }
    Page:
      type: object
      properties:
        cursor: { type: string, nullable: true }
        has_more: { type: boolean }
...
    Message:
      type: object
      properties:
        id: { type: string, format: uuid }
        sender:
          type: object
          properties:
            id: { type: string, format: uuid }
            display_name: { type: string }
            role: { type: string, nullable: true, enum: [parent, educator, executive, admin] }
        kind: { type: string, enum: [text, voice, file] }
        body: { type: string, nullable: true, description: Null for members once hidden; oversight keeps the text }
        duration_seconds: { type: integer, nullable: true }
        created_at: { type: string, format: date-time }
        hidden:
          type: object
          nullable: true
          properties:
            by_id: { type: string, format: uuid }
            by_name: { type: string }
            at: { type: string, format: date-time }
        report:
          type: object
          nullable: true
          description: Only sent to oversight callers while the report is open
          properties:
            id: { type: string, format: uuid }
            reporter_id: { type: string, format: uuid }
            reporter_name: { type: string }
            reason: { type: string }
    ReviewPost:
      type: object
      properties:
        id: { type: string, format: uuid }
        album:
          type: object
          properties:
            id: { type: string, format: uuid }
            title: { type: string }
            group_name: { type: string, nullable: true }
            moderation_mode: { type: string, enum: [publish_then_moderate, approve_before_publish] }
        author_id: { type: string, format: uuid }
        author_name: { type: string }
        created_at: { type: string, format: date-time }
        media_count: { type: integer }
        media_kind: { type: string, enum: [photo, video, banner] }
        thumbnail_url: { type: string, nullable: true }
        tags:
          type: array
          items:
            type: object
            properties:
              child_id: { type: string, format: uuid }
              full_name: { type: string }
              image_rights_level: { type: string, enum: [allowed, app_only, not_allowed], description: The child's current level, resolved at read time }
        moderation_status: { type: string, enum: [pending, published, hidden] }
        is_blocked: { type: boolean, description: Hidden by a consent downgrade after publication }
...
  responses:
...
    Forbidden:
      description: Authenticated, but outside the caller's permission scope
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
paths:
...
  /conversations/{conversationId}:
    get:
      summary: A conversation's header — title, membership and member labels (never a phone number, MSG-06)
      parameters:
        - $ref: '#/components/parameters/conversationId'
      responses:
        '200': { description: OK }
        '403': { $ref: '#/components/responses/Forbidden' }
  /conversations/{conversationId}/messages:
    get:
      summary: The thread, oldest first. A non-member's read is oversight and is audit-logged before the rows are returned.
      parameters:
        - $ref: '#/components/parameters/conversationId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Message' } }
                  page: { $ref: '#/components/schemas/Page' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      summary: Send a text message
      parameters:
        - $ref: '#/components/parameters/conversationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kind, body]
              properties:
                kind: { type: string, enum: [text] }
                body: { type: string, maxLength: 4000 }
      responses:
        '201':
          description: Created
          content: { application/json: { schema: { $ref: '#/components/schemas/Message' } } }
  /messages/{messageId}/hide:
    post:
      summary: Hide a message (oversight only). Reversible, audit-logged, closes any open report; never a deletion.
      parameters:
        - name: messageId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Hidden }
        '403': { $ref: '#/components/responses/Forbidden' }
  /message-reports/{reportId}/dismiss:
    post:
      summary: Close a report without hiding the message (oversight only, audit-logged)
      parameters:
        - name: reportId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Dismissed }
        '403': { $ref: '#/components/responses/Forbidden' }
  /memories/review-queue:
    get:
      summary: Posts awaiting moderation (EXEC-M-04) — pending, plus published posts auto-hidden by a consent downgrade
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  moderation_mode:
                    type: string
                    nullable: true
                    enum: [publish_then_moderate, approve_before_publish]
                    description: The one mode every album in scope runs under; null when unset or mixed (open decision #3)
                  data: { type: array, items: { $ref: '#/components/schemas/ReviewPost' } }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /memories/posts/{postId}/approve:
    post:
      summary: Publish a pending post, or re-publish a blocked one. Re-checks every tagged child's current image rights (WAL-06).
      parameters:
        - $ref: '#/components/parameters/postId'
      responses:
        '204': { description: Published }
        '422':
          description: memories.consent_blocked — `details.child_ids` names the children whose guardians said no
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  /memories/posts/{postId}/hide:
    post:
      summary: Hide a post. Never a deletion.
      parameters:
        - $ref: '#/components/parameters/postId'
      responses:
        '204': { description: Hidden }
...